Move validate out of the helper package 🌿
